chirpstack-docker/Automated_Metrics/requirements.txt
paho-mqtt
influxdb
python-dateutil